all helper functions have been moved to the module utils.py the remaining modules have been removed
