+++ data/exchange_rates.json
@@ -1,10 +1,10 @@
 {
   "USD": {
-    "rate": 86.3829,
-    "date": "Fri, 14 Mar 2025 00:00:02 +0000"
+    "rate": 85.4998,
+    "date": "Mon, 17 Mar 2025 00:00:01 +0000"
   },
   "EUR": {
-    "rate": 93.7906,
-    "date": "Fri, 14 Mar 2025 00:00:02 +0000"
+    "rate": 93.0497,
+    "date": "Mon, 17 Mar 2025 00:00:01 +0000"
   }
 }
